chore: clean up dead code, fix clippy warnings (#27)
- Replace manual Default impls with #[derive(Default)] for IaConfig, LoggingConfig
- Use struct initializer with ..Default::default() to avoid field_reassign_with_default
- Remove unnecessary u64 casts in disk_pool.rs
- Add OnItemStartFn/OnItemCompleteFn type aliases to reduce type_complexity
- Update --tui references to --dashboard in README.md and CHANGELOG.md

+++ ia-core/src/disk_pool.rs
@@ -186,8 +186,8 @@ fn disk_space(path: &Path) -> Result<(u64, u64)> {
             )));
         }
 
-        let free = stat.f_bavail as u64 * stat.f_frsize as u64;
-        let total = stat.f_blocks as u64 * stat.f_frsize as u64;
+        let free = stat.f_bavail as u64 * stat.f_frsize;
+        let total = stat.f_blocks as u64 * stat.f_frsize;
         Ok((free, total))
     }
